Added new package for auto commands and moved Auto_DriveStraight
